* no longer listing the logs in the top-level md-status handler (where all managed    domains are listed). Getting the status of a particular domain has the log).  * fixing some signedness warnings in clang 11.

+++ src/md_store.c
@@ -59,9 +59,9 @@ static const char *GROUP_NAME[] = {
     NULL
 };
 
-const char *md_store_group_name(int group)
+const char *md_store_group_name(unsigned int group)
 {
-    if ((size_t)group < sizeof(GROUP_NAME)/sizeof(GROUP_NAME[0])) {
+    if (group < sizeof(GROUP_NAME)/sizeof(GROUP_NAME[0])) {
         return GROUP_NAME[group];
     }
     return "UNKNOWN";
